virt/common: Add support for shared virtiofs mount
Create a "shared" directory under cuttlefish_runtime.X which is
unconditionally shared with the guest. This virtiofs filesystem is
mounted at /mnt/vendor/shared and is accessible only to root. This
is similar to how 'persist' works on Pixel.

Implementing this for QEMU is postponed for now; the feature was only
added in QEMU 5.0 and requires a new subprocess called virtiofsd to
be started; we can revisit this when QEMU 5.0 has rolled out to more
places. Added "nofail" to the fstab to accommodate this.

Bug: 161843089
Change-Id: Id69404f0ac55ec5f1aff11412e648fb5260a3f2c